Reorganized sections
